- Added support for non-compressed update opcodes - Changed UPDATETYPE_VALUES so the update values gets changed in the object. - Updated misc other stuff to the UPDATETYPE_VALUES code
git-svn-id: https://pseuwow.svn.sourceforge.net/svnroot/pseuwow@66 fda9a0ab-4e26-0410-ba80-894f06a4b443

+++ src/Client/World/Object.h
@@ -65,7 +65,7 @@ public:
     {
         _floatvalues[ index ] = value;
     }
-    inline void SetUInt32Value( uint16 index, float value )
+    inline void SetUInt32Value( uint16 index, uint32 value )
     {
         _uint32values[ index ] = value;
     }
